Move agency methods from CounterpartyServices to new AgencyServices
