+++ Api/Controllers/AgentControllers/CounterpartiesController.cs
@@ -1,11 +1,8 @@
-using System.Collections.Generic;
 using System.Net;
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
-using HappyTravel.Edo.Api.Filters.Authorization.AgentExistingFilters;
 using HappyTravel.Edo.Api.Filters.Authorization.InAgencyPermissionFilters;
 using HappyTravel.Edo.Api.Infrastructure;
-using HappyTravel.Edo.Api.Models.Agencies;
 using HappyTravel.Edo.Api.Models.Agents;
 using HappyTravel.Edo.Api.Services.Agents;
 using HappyTravel.Edo.Api.Services.Files;
@@ -48,37 +45,6 @@ namespace HappyTravel.Edo.Api.Controllers.AgentControllers
         //        ? (IActionResult) NoContent()
         //        : BadRequest(ProblemDetailsBuilder.Build(error));
         //}
-
-
-        /// <summary>
-        ///     Gets agency.
-        /// </summary>
-        /// <param name="agencyId">Agency Id.</param>
-        /// <returns></returns>
-        [HttpGet("agencies/{agencyId}")]
-        [ProducesResponseType((int) HttpStatusCode.NoContent)]
-        [ProducesResponseType(typeof(ProblemDetails), (int) HttpStatusCode.BadRequest)]
-        [AgentRequired]
-        public async Task<IActionResult> GetAgency(int agencyId)
-        {
-            var (_, isFailure, agency, error) = await _counterpartyService.GetAgency(agencyId, await _agentContextService.GetAgent());
-
-            if (isFailure)
-                return BadRequest(ProblemDetailsBuilder.Build(error));
-
-            return Ok(agency);
-        }
-
-
-        /// <summary>
-        ///     Gets child agencies.
-        /// </summary>
-        /// <returns></returns>
-        [HttpGet("agency/child-agencies")]
-        [ProducesResponseType(typeof(List<AgencyInfo>), (int)HttpStatusCode.OK)]
-        [InAgencyPermissions(InAgencyPermissions.ObserveChildAgencies)]
-        public async Task<IActionResult> GetChildAgencies()
-            => Ok(await _counterpartyService.GetChildAgencies(await _agentContextService.GetAgent()));
 
 
         /// <summary>
