+++ Api/Services/Agents/CounterpartyService.cs
@@ -1,10 +1,8 @@
-using System.Collections.Generic;
+using System;
 using System.Linq;
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
-using HappyTravel.Edo.Api.AdministratorServices;
 using HappyTravel.Edo.Api.Infrastructure;
-using HappyTravel.Edo.Api.Models.Agencies;
 using HappyTravel.Edo.Api.Models.Agents;
 using HappyTravel.Edo.Common.Enums;
 using HappyTravel.Edo.Data;
@@ -17,14 +15,10 @@ namespace HappyTravel.Edo.Api.Services.Agents
     public class CounterpartyService : ICounterpartyService
     {
         public CounterpartyService(EdoContext context,
-            IAgentService agentService,
-            IDateTimeProvider dateTimeProvider,
-            ICounterpartyManagementService counterpartyManagementService)
+            IDateTimeProvider dateTimeProvider)
         {
             _context = context;
-            _agentService = agentService;
             _dateTimeProvider = dateTimeProvider;
-            _counterpartyManagementService = counterpartyManagementService;
         }
 
 
@@ -138,18 +132,6 @@ namespace HappyTravel.Edo.Api.Services.Agents
         //}
 
 
-        public async Task<Result<AgencyInfo>> GetAgency(int agencyId, AgentContext agent)
-        {
-            var agentRelations = await _agentService.GetAgentRelations(agent);
-            if (agentRelations.All(r => r.AgencyId != agencyId))
-                return Result.Failure<AgencyInfo>("The agent is not affiliated with agency");
-
-            var agency = await _context.Agencies.SingleAsync(a => a.Id == agencyId);
-
-            return Result.Success(new AgencyInfo(agency.Name, agency.Id));
-        }
-
-
         public Task<Agency> GetDefaultAgency(int counterpartyId)
             => _context.Agencies
                 .SingleAsync(a => a.CounterpartyId == counterpartyId && a.ParentId == null);
@@ -159,9 +141,6 @@ namespace HappyTravel.Edo.Api.Services.Agents
         {
             return await GetCounterpartyInfo(counterpartyId, languageCode);
         }
-
-
-        public Task<List<AgencyInfo>> GetChildAgencies(AgentContext agent) => _counterpartyManagementService.GetChildAgencies(agent.AgencyId);
 
 
         private async Task<Result<CounterpartyInfo>> GetCounterpartyInfo(int counterpartyId, string languageCode = LocalizationHelper.DefaultLanguageCode )
@@ -196,10 +175,8 @@ namespace HappyTravel.Edo.Api.Services.Agents
                 result.Counterparty.IsContractUploaded));
         }
 
-
-        private readonly IAgentService _agentService;
+        
         private readonly EdoContext _context;
         private readonly IDateTimeProvider _dateTimeProvider;
-        private readonly ICounterpartyManagementService _counterpartyManagementService;
     }
 }
