+++ Api/Services/Agents/ICounterpartyService.cs
@@ -1,8 +1,6 @@
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
 using HappyTravel.Edo.Api.Infrastructure;
-using HappyTravel.Edo.Api.Models.Agencies;
 using HappyTravel.Edo.Api.Models.Agents;
 using HappyTravel.Edo.Data.Agents;
 
@@ -16,10 +14,6 @@ namespace HappyTravel.Edo.Api.Services.Agents
 
         //Task<Result<Agency>> AddAgency(int counterpartyId, AgencyInfo agency);
 
-        Task<Result<AgencyInfo>> GetAgency(int agencyId, AgentContext agent);
-
         Task<Agency> GetDefaultAgency(int counterpartyId);
-
-        Task<List<AgencyInfo>> GetChildAgencies(AgentContext agent);
     }
 }
