+++ HappyTravel.Edo.UnitTests/Tests/AdministratorServices/CounterpartyManagementServiceTests/ActivationTests.cs
@@ -67,7 +67,7 @@ namespace HappyTravel.Edo.UnitTests.Tests.AdministratorServices.CounterpartyMana
 
 
         [Fact]
-        public async Task Activation_of_counterparty_should_activate_agencies()
+        public async Task Activation_of_counterparty_should_not_activate_agencies()
         {
             var context = _mockCreationHelper.GetContextMock().Object;
             var counterpartyManagementService = _mockCreationHelper.GetCounterpartyManagementService(context);
@@ -76,7 +76,7 @@ namespace HappyTravel.Edo.UnitTests.Tests.AdministratorServices.CounterpartyMana
 
             var agencies = context.Agencies.Where(ag => ag.CounterpartyId == 2).ToList();
             Assert.False(isFailure);
-            Assert.True(agencies.All(ag => ag.IsActive));
+            Assert.DoesNotContain(agencies.Where(a => a.Id == 3 || a.Id == 5), ag => ag.IsActive);
         }
 
 
@@ -98,9 +98,9 @@ namespace HappyTravel.Edo.UnitTests.Tests.AdministratorServices.CounterpartyMana
         public async Task Activation_of_active_agency_should_succeed()
         {
             var context = _mockCreationHelper.GetContextMock().Object;
-            var counterpartyManagementService = _mockCreationHelper.GetCounterpartyManagementService(context);
+            var agencyManagementService = _mockCreationHelper.GetAgencyManagementService(context);
 
-            var (_, isFailure, error) = await counterpartyManagementService.ActivateAgency(1, "Test reason");
+            var (_, isFailure, error) = await agencyManagementService.ActivateAgency(1, "Test reason");
 
             Assert.False(isFailure);
         }
@@ -110,9 +110,9 @@ namespace HappyTravel.Edo.UnitTests.Tests.AdministratorServices.CounterpartyMana
         public async Task Activation_of_not_existing_agency_should_fail()
         {
             var context = _mockCreationHelper.GetContextMock();
-            var counterpartyManagementService = _mockCreationHelper.GetCounterpartyManagementService(context.Object);
+            var agencyManagementService = _mockCreationHelper.GetAgencyManagementService(context.Object);
 
-            var (_, isFailure, error) = await counterpartyManagementService.ActivateAgency(105, "Test reason");
+            var (_, isFailure, error) = await agencyManagementService.ActivateAgency(105, "Test reason");
 
             Assert.True(isFailure);
         }
@@ -122,9 +122,9 @@ namespace HappyTravel.Edo.UnitTests.Tests.AdministratorServices.CounterpartyMana
         public async Task Activation_of_agency_with_empty_reason_should_fail()
         {
             var context = _mockCreationHelper.GetContextMock();
-            var counterpartyManagementService = _mockCreationHelper.GetCounterpartyManagementService(context.Object);
+            var agencyManagementService = _mockCreationHelper.GetAgencyManagementService(context.Object);
 
-            var (_, isFailure, error) = await counterpartyManagementService.ActivateAgency(3, null);
+            var (_, isFailure, error) = await agencyManagementService.ActivateAgency(3, null);
 
             Assert.True(isFailure);
         }
@@ -134,9 +134,9 @@ namespace HappyTravel.Edo.UnitTests.Tests.AdministratorServices.CounterpartyMana
         public async Task Activation_of_agency_should_activate()
         {
             var context = _mockCreationHelper.GetContextMock().Object;
-            var counterpartyManagementService = _mockCreationHelper.GetCounterpartyManagementService(context);
+            var agencyManagementService = _mockCreationHelper.GetAgencyManagementService(context);
 
-            var (_, isFailure, error) = await counterpartyManagementService.ActivateAgency(3, "Test reason");
+            var (_, isFailure, error) = await agencyManagementService.ActivateAgency(3, "Test reason");
 
             var agency = context.Agencies.Single(ag => ag.Id == 3);
             Assert.False(isFailure);
@@ -148,9 +148,9 @@ namespace HappyTravel.Edo.UnitTests.Tests.AdministratorServices.CounterpartyMana
         public async Task Activation_of_agency_should_activate_agents_relations()
         {
             var context = _mockCreationHelper.GetContextMock().Object;
-            var counterpartyManagementService = _mockCreationHelper.GetCounterpartyManagementService(context);
+            var agencyManagementService = _mockCreationHelper.GetAgencyManagementService(context);
 
-            var (_, isFailure, error) = await counterpartyManagementService.ActivateAgency(3, "Test reason");
+            var (_, isFailure, error) = await agencyManagementService.ActivateAgency(3, "Test reason");
 
             var agentIds = new List<int> { 5, 6 };
             var agentAgencyRelations = context.AgentAgencyRelations.Where(ag => agentIds.Contains(ag.AgentId)).ToList();
@@ -163,9 +163,9 @@ namespace HappyTravel.Edo.UnitTests.Tests.AdministratorServices.CounterpartyMana
         public async Task Activation_of_agency_should_activate_accounts()
         {
             var context = _mockCreationHelper.GetContextMock().Object;
-            var counterpartyManagementService = _mockCreationHelper.GetCounterpartyManagementService(context);
+            var agencyManagementService = _mockCreationHelper.GetAgencyManagementService(context);
 
-            var (_, isFailure, error) = await counterpartyManagementService.ActivateAgency(3, "Test reason");
+            var (_, isFailure, error) = await agencyManagementService.ActivateAgency(3, "Test reason");
 
             var account = context.AgencyAccounts.Single(p => p.Id == 2);
             Assert.False(isFailure);
@@ -177,9 +177,9 @@ namespace HappyTravel.Edo.UnitTests.Tests.AdministratorServices.CounterpartyMana
         public async Task Activation_of_agency_should_activate_child_agencies()
         {
             var context = _mockCreationHelper.GetContextMock().Object;
-            var counterpartyManagementService = _mockCreationHelper.GetCounterpartyManagementService(context);
+            var agencyManagementService = _mockCreationHelper.GetAgencyManagementService(context);
 
-            var (_, isFailure, error) = await counterpartyManagementService.ActivateAgency(3, "Test reason.");
+            var (_, isFailure, error) = await agencyManagementService.ActivateAgency(3, "Test reason.");
 
             var childAgency = context.Agencies.Single(ag => ag.Id == 5);
             Assert.False(isFailure);
@@ -188,16 +188,16 @@ namespace HappyTravel.Edo.UnitTests.Tests.AdministratorServices.CounterpartyMana
 
 
         [Fact]
-        public async Task Activation_of_default_agency_should_activate_counterparty()
+        public async Task Activation_of_default_agency_should_not_activate_counterparty()
         {
             var context = _mockCreationHelper.GetContextMock().Object;
-            var counterpartyManagementService = _mockCreationHelper.GetCounterpartyManagementService(context);
+            var agencyManagementService = _mockCreationHelper.GetAgencyManagementService(context);
 
-            var (_, isFailure, error) = await counterpartyManagementService.ActivateAgency(3, "Test reason");
+            var (_, isFailure, error) = await agencyManagementService.ActivateAgency(3, "Test reason");
 
             var counterparty = context.Counterparties.Single(c => c.Id == 2);
             Assert.False(isFailure);
-            Assert.True(counterparty.IsActive);
+            Assert.False(counterparty.IsActive);
         }
